minor redux refactoring
