+++ packages/quizzes-dashboard/store/editor/options/optionReducer.ts
@@ -1,10 +1,7 @@
 import { action, Option } from "../../../types/NormalizedQuiz"
 import { createReducer } from "typesafe-actions"
-import {
-  initializedEditor,
-  editedOptionTitle,
-  editedOptionCorrectnes,
-} from "../../editor/editorActions"
+import { editedOptionTitle, editedOptionCorrectnes } from "./optionActions"
+import { initializedEditor } from "../editorActions"
 
 export const optionReducer = createReducer<
   { [optionId: string]: Option },
@@ -14,11 +11,13 @@ export const optionReducer = createReducer<
     initializedEditor,
     (_state, action) => action.payload.quiz.options ?? {},
   )
+
   .handleAction(editedOptionTitle, (state, action) => {
     let newState = state
     newState[action.payload.optionId].title = action.payload.newTitle
     return newState
   })
+
   .handleAction(editedOptionCorrectnes, (state, action) => {
     let newState = state
     newState[action.payload.optionId].correct = !newState[
